# ENH: some micro-optimisations for differential_evolution