Rationale:
`_select_samples` and `_ensure_constraint` are called repeatedly and they need to be efficient (hundreds of thousands of times). This PR makes some microoptimisations to make both those methods faster.
The changes to _ensure_constraint are based on:

```
a = rng.uniform(size=1000, low=-0.1, high=1.1)
msk = np.bitwise_or(a < 0, a > 1)
%timeit np.where((a < 0) | (a > 1))
%timeit np.bitwise_or(a < 0, a > 1)
%timeit rng.uniform(size=0)
%timeit np.count_nonzero(msk)
3.29 µs ± 6.57 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
2.19 µs ± 6.41 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
1.19 µs ± 35.5 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
235 ns ± 1.61 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```

i.e. `bitwise_or` is faster than `np.where`, and that `rng.uniform(size=0)` has a cost. Most of the time `msk` has zero entries, so it's quicker to check how many entries that do need replacing, and only if there are entries is it worth calling the random number generator.

The changes to `_select_samples` are based on:

``` python
num_population_members = 200
rng = np.random.default_rng()
pool = np.arange(num_population_members)

def select_samples(candidate, number_samples):
    pool = np.arange(num_population_members)
    rng.shuffle(pool)

    idxs = []
    while len(idxs) < number_samples and len(pool) > 0:
        idx = pool[0]
        pool = pool[1:]
        if idx != candidate:
            idxs.append(idx)

    return idxs


def select_samples2(candidate, number_samples, pool):
    rng.shuffle(pool) # also tried rng.choice, but that's slower

    idxs = pool[:number_samples + 1]
    return idxs[idxs != candidate][:number_samples]

%timeit select_samples(5, 5)
%timeit select_samples2(5, 5, pool)
3.71 µs ± 10.3 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
3.23 µs ± 27 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```