@andyfaff 2024-05-20
The tests had to be modified to account for the changes.

In the custom strategy function test the minimisation pathway for the custom strategy is slightly different to the inbuilt best1bin strategy. In this PR `_select_samples` reshuffles an array that's been added as an attribute to the class. Previously, and in the custom strategy function, the shuffle occurred on a sorted array.  Thus it didn't make sense to check for bitwise consistency, a smoke test is sufficient.