@andyfaff 2024-05-20
Before last commit
```
%lprun -f DifferentialEvolutionSolver._scale_parameters differential_evolution(rosen, bounds)
Timer unit: 1e-09 s

Total time: 0.852551 s
File: /Users/andrew/Documents/Andy/programming/scipy/build-install/lib/python3.12/site-packages/scipy/optimize/_differentialevolution.py
Function: _scale_parameters at line 1648

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  1648                                               def _scale_parameters(self, trial):
  1649                                                   """Scale from a number between 0 and 1 to parameters."""
  1650                                                   # trial either has shape (N, ) or (L, N), where L is the number of
  1651                                                   # solutions being scaled
  1652    151003  195571000.0   1295.1     22.9          scaled = self.__scale_arg1 + (trial - 0.5) * self.__scale_arg2
  1653    151003  619885000.0   4105.1     72.7          if np.any(self.integrality):
  1654                                                       i = np.broadcast_to(self.integrality, scaled.shape)
  1655                                                       scaled[i] = np.round(scaled[i])
  1656    151003   37095000.0    245.7      4.4          return scaled
```

After last commit
```
%lprun -f DifferentialEvolutionSolver._scale_parameters differential_evolution(rosen, bounds)
Timer unit: 1e-09 s

Total time: 0.361521 s
File: /Users/andrew/Documents/Andy/programming/scipy/build-install/lib/python3.12/site-packages/scipy/optimize/_differentialevolution.py
Function: _scale_parameters at line 1648

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
  1648                                               def _scale_parameters(self, trial):
  1649                                                   """Scale from a number between 0 and 1 to parameters."""
  1650                                                   # trial either has shape (N, ) or (L, N), where L is the number of
  1651                                                   # solutions being scaled
  1652    151003  193316000.0   1280.2     53.5          scaled = self.__scale_arg1 + (trial - 0.5) * self.__scale_arg2
  1653    151003  134241000.0    889.0     37.1          if np.count_nonzero(self.integrality):
  1654                                                       i = np.broadcast_to(self.integrality, scaled.shape)
  1655                                                       scaled[i] = np.round(scaled[i])
  1656    151003   33964000.0    224.9      9.4          return scaled
```